Add a simple `PrometheusRule` to check successful Frontend deployment
This commit adds two simple `Prometheus` rules. The first checks that there are pods running Frontend, whilst the second checks that Frontend has been successfully deployed.

This second rule uses the [`kube_deployment_status_condition` deployment metric](https://github.com/kubernetes/kube-state-metrics/blob/master/docs/deployment-metrics.md) to check that, for each namespace Frontend is deployed into, the `frontend` deployment returns `true` for both being `Available` and `Progressing`.

Evaluating the status of a deployment within Kubernetes does not seem to be as trivial within Prometheus as one might expect. [According to the docs](https://kubernetes.io/docs/concepts/workloads/controllers/_print/#complete-deployment), a deployment which is `Available` means that the minimum number of replicas have been satisfied, as specified in a Deployment spec; it doesn't necessarily mean that the entire deployment is complete. A deployment which is `Progressing` means either that the deployment is creating pods etc and _is_ actually in progress, or that it has completed successfully, specifying the `Reason` as `NewReplicaSetAvailable`.

Unfortunately, `Reason` does not appear to be available within the `kube_deployment_status_condition` deployment metric and as such we can't use it to properly check whether a deployment was successful. For this reason, this commit is only checking for any deployments of Frontend where `Progressing` is false and `ReplicaFailure` is true.

For the deployment rule, it's worth noting the reason behind this rule is to check that Prometheus fires the alert when appropriate (when a deployment of frontend fails) and that this firing alert shows up in AlertManager; it isn't intended as a best-practice example of what a good `PrometheusRule` looks like.