Add `cluster-security` Helm chart
This commit adds a new Helm chart, `cluster-security`. This chart adds a baseline `PodSecurityPolicy` to the cluster, [in line with the recommendations on the docs](https://kubernetes.io/docs/concepts/policy/pod-security-policy/#recommended-practice). With this cluster-wide policy, containers running as root and with privilege escalation are blocked, whilst containers are also required to run with a read-only root filesystem. This policy is likely to be refined over time and does not represent the final state of the policy in its current form.

This policy applies to any authenticated users and is enforced when modifying the cluster, via the use of Helm and also directly via `kubectl`.