Set RAILS_ENV for Static
This commit sets the `RAILS_ENV` var for Static from a new `appRailsEnv` property in the chart `values.yaml`. Previously `govukEnvironment` was not being set and as a result would have defaulted Static to `development`, which isn't what we want. Intoducing `appRailsEnv` also makes this chart consistent with the others.