Configure ArgoCD to deploy apps
This adds two charts:

* argocd-config, responsible for configuring ArgoCD, including
  projects, repos and repo creds. This is installed by cluster-services
  in Terraform.
* argocd-apps, responsible for creating Argo Application resources,
  and defining the env-specific values for the Helm charts installed
  by Argo (i.e., this is where environment image tags are stored).

Thus far we've added only cluster-services, signon, and publisher
to Argo, but this should suffice as a template for other applications.

Any changes to an Argo-managed Helm chart OR the ArgoCD Application
will trigger a sync (deploy) of the application.

We use two helm charts (argocd-config & argocd-apps) to avoid a circular
dependency where an ArgoCD Application tries to sync itself - this is
something to avoid as in my experience it leads to unresolvable states
where (e.g.) the app is constantly syncing.

We benefit from packaging the argocd-config chart as a Helm chart since
this enables us to install the chart in the cluster-services Terraform
deployment in govuk-infrastructure consistent with other K8s resources.