Make db migration job name statically defined
Using a Helm function to give a unique name to a resource
causes the resource to be recreated on each commit to the
repo due to Argo's sync policy.

In effect, publisher's db migration would run with each commit
to the govuk-helm-charts repo, since the result of helm template
would yield a difference between the live and desired configuration,
causing Argo to recreate the job.

By removing the dynamic naming from the job, we can run db migrations
when publisher is deployed using a PreSync Argo hook.

Co-Authored-By: Frederic <frederic.francois@digital.cabinet-office.gov.uk>