Send a Webhook to Argo Events after apps are deployed
This defines the ArgoCD Notifications configuration necessary
to handle the on-deployed subscription on Argo Applications
(see prior commit).

Once an application is synced and healthy, a webhook is sent to
the Argo Events EventSource called sync-notification at the
/argocd endpoint. The webhook contains the necessary context
(such as app, state, commit, slackChannel) to run any post-sync
jobs as necessary for the app (i.e. smoke tests, slack alerts,
release promotion).

See ArgoCD Notifications docs for more detail on this config:
https://argocd-notifications.readthedocs.io/en/stable.

Co-Authored-By: Frederic <frederic.francois@digital.cabinet-office.gov.uk>