Add Argo Events configuration
This adds the necessary config (eventbus, sensor, source, service account)
for Argo Events to receive webhooks from ArgoCD Notifications when
an application is deployed, and trigger an Argo Workflow called
post-sync (not defined in this commit).

See the documentation for Argo Events at
https://argoproj.github.io/argo-events.

The EventSource 'sync-notification' places sync events on the eventbus,
which are then captured by the 'sync-notification' Sensor, which
triggers the post-sync workflow.

The Sensor passes the sync context (app, commit, slack-channel) which
will be used by the Argo Workflow (named post-sync) to run jobs such
as smoke tests after a deploy.

Co-Authored-By: Frederic <frederic.francois@digital.cabinet-office.gov.uk>