Add Argo Workflows for post-sync jobs
This adds a post-sync WorkflowTemplate that is triggered
by Argo Events and runs the smoke tests for an app and reports
the results to Slack.

Later we can add another promote-release WorkflowTemplate, called
from the post-sync workflow after smoke tests have passed.

This commit adds the argo-workflows Helm chart, which we

Co-Authored-By: Frederic <frederic.francois@digital.cabinet-office.gov.uk>