Use unique but human-friendly names for ALBs created by Ingresses.
By default, aws-lb-controller creates load balancers with random strings
as names. We want to be able to look up the LB for a given Ingress
easily and vice versa. We previously acheived this using fixed names,
but these would clash when deploying the same apps into more than one
namespace for testing.

Include the namespace in the ALB name so that the names are unique while
also being human-friendly.

This doesn't cater for overriding the ALB name via Helm values, but
I can't think of any valid reason for wanting to do that at this stage.
We can easily add a value for it later if we need it.

Tested: charts produce valid kube yaml and apps come up successfully on
install.