Move common env vars from `frontend` to a shared ConfigMap.
We introduce a new govuk-apps-env ConfigMap to represent the set of
environment variables which are common to all the GOV.UK apps
(plus some which will have a couple of overrides in the draft apps - but
that's ok).

Also:
- Remove some unused/unnecessary/inappropriate env vars.
- Rename configmap.yaml to nginx-configmap.yaml for clarity.

The following are the same for all apps in a given environment, so
they're in the shared ConfigMap:

- ASSET_HOST
- GOVUK_APP_DOMAIN (not actually used anyway because we're setting all
  the PLEK_SERVICE_*_URI explicitly, but Plek requires it for now)
- GOVUK_APP_DOMAIN_EXTERNAL
- GOVUK_ASSET_ROOT
- GOVUK_ENVIRONMENT
- GOVUK_WEBSITE_ROOT
- PLEK_SERVICE_*_URI (except for draft apps, which will override one or
  two of these)

The following are unnecessary or otherwise don't belong in the app chart:

- GOVUK_APP_NAME: this is an inherent property of the app and doesn't
  need to be overridden, therefore we're setting it in the Dockerfile.
  It should probably be removed entirely at some point.
- GOVUK_APP_TYPE: used only in govuk-puppet, not in the app
- DEFAULT_TTL: only relevant to Content Store.
- GOVUK_CONTENT_SCHEMAS_PATH: only relevant to Publishing API and test
  runners (see
  https://drive.google.com/open?id=1nxJzdbTK__jWDFjUj9C5sH4yP0caIHhXp6fx1yKGHwE)
- RAILS_ENV: this is a property of the container image; it would never
  make sense to run a production image in dev mode - you'd need a
  different image that includes the dev dependencies.
- RAILS_LOG_TO_STDOUT: this always has to be set when running in a
  container, so it belongs in the Dockerfile.

Normally a change like this would merit a major version bump, but I've
made it a minor because the major is still 0 and we're the only user. It
seemed wrong to go to 1.0 at this point as that would be (more) misleading.

Tested: frontend still works in the test cluster with these changes.

Trello: https://trello.com/c/h6lGL53t/706