Tweak liveness/readiness/startup probes to reduce startup delay.
Use a startup probe for the Rails container to avoid an unnecessarily
long delay on startup.

Also tweak some of the existing timeouts and clear up some of the
boilerplate with YAML anchors. YAML anchors are usually best avoided in
Helm charts but I think in this case the DRYness makes the config a bit
more robust.