Set govukStack correctly for the integration environment.
govukStack is a terrible wart, but we can't quite get rid of it just
yet. We need to change the hostnames of MongoDB cluster nodes in the
test account from `pink` to `blue` first, so until we get around to
doing that it'll stick around, complicating up the Helm charts.