Avoid duplicate lint checks on first-party PRs.
Alternatively, we could just remove `pull_requests` from the `on`
predicate. That'd mean we don't run checks automatically on third-party
PRs, which has pros and cons: better for security, bit inconvenient for
accepting third-party contributions (but this happens so rarely that
we're probably talking about a handful of minutes a year).