Run checks on push only, no longer on PR.
The GitHub Slack bot doesn't play nice with skipped checks: it says
things like "1/2 successful checks" when one of the checks is skipped.
So let's just get rid of the checks on PR.

The (theoretical) disadvantage of this approach is that check won't run
automatically on third-party PRs, but we've literally never had one of
these in the entire history of this repo or its predecessor,
alphagov/govuk-aws.