Use YAML for Helm values instead of setting key/value params.
This means we have the same syntax when passing Helm values via Argo as
when deploying using Helm directly, which makes testing easier as well
as making the config more readable.

Tested: tried it on the test cluster by editing app/argocd-apps to point
at the branch. The Argo apps seem to work as expected.