Use the external domain suffix for Signon.
PLEK_SERVICE_*_URI overrides both the "internal" and the "external"
URL for a given service. Plek doesn't currently provide a way to
override them separately.

This workaround trombones the internal traffic to the edge and back
(which is bad for reliability), but it at least keeps things working for
now until we fix up Plek so that we can do avoid the trombone routing.

The cleaner medium-term solution is probably to add a "use plain http for
internal stuff" option to Plek and then let it do its usual
thing instead of explicitly overriding all the service URLs, i.e. let it
append GOVUK_APP_DOMAIN or GOVUK_APP_DOMAIN_EXTERNAL depending on
whether the app wants to construct an internal or external URL.