Replace Kuberentes Jobs during Sync rather than upgrading in-place
Jobs are immutable so our sync policy should use `kubectl replace`
rather than `kubectl apply`.

Docs: https://argo-cd.readthedocs.io/en/stable/user-guide/sync-options/#replace-resource-instead-of-applying-changes

This is configured at the individual resource level.

The motivation for this change is that DB migration jobs can fail
or need to be re-run, and we'd like to ensure that we can re-sync
those objects successfully.