Split up AWS Secrets Manager secrets for Publisher.
Publisher secrets were being stored in a single AWS Secrets Manager
secret. This was a temporary hack to get the app running quickly, but
would prevent us from being able to rotate and manage the credentials
contained in that all-in-one Secret.

Split the logical secrets up into separate Secrets Manager secrets so
that we don't break auditability and automatability.

For now, we're putting the ExternalSecrets definitions in the
govuk-apps-conf Helm chart, for want of a better place. This isn't
ideal; they really belong in the apps' own Helm charts (behind an "if
.Values.enableExternalSecrets") but this won't be feasible until we
refactor our shared govuk-rails-app Helm chart into per-app charts with
library chart(s) - i.e. shift from inheritance to composition (as
@bilbof eruditely put it).

Some of the ExternalSecrets definitions require the use of Go templates
in order to generate (for example) database connection strings (see
https://external-secrets.io/guides-templating/). The MongoDB connection
string here is one example but there will be others.  These templates
(which are Go templates but not Helm templates and are processed by the
ExternalSecrets operator) are kept in separate files and included
verbatim in the Helm templates in order to avoid horrible
double-escaping like this:

```dont-try-this-at-home-kids
{{ "{{" }} .username | toString {{ "}}" }}:{{ "{{" }} .password | toString {{ "}}" }}@{{ "{{" }} .hostname | toString {{ "}}" }}
```
🤮🤮🤮

Alternatives considered:

Thought about passing the ExternalSecrets definitions into the shared
chart as Helm values, but given that ExternalSecrets can contain
(non-Helm) Go templates of their own, the number of levels of string
templating and YAML embedding involved would be most unwise (see above -
it'd be even worse than that 😱).

Generating the ExternalSecrets definitions in a Helm template in
govuk-rails-app was another option. This would have avoided the very
worst of the double-escaping but would have involved implementing a fair
bit of logic inside a Helm template and defining a nontrivial structure
for passing in the necessary information from values.yaml about how to
construct the ExternalSecrets. Still a lot of unwarranted layers of
templating and abstraction, i.e. complexity that doesn't seem worth it
compared to the chosen approach above.

Also considered using a separate ConfigMap for templating the MongoDB
connection string instead of using ExternalSecrets Operator's
templating, but concluded that this only moves the templating problem
rather than simplifying it, merely resulting in an extra level of
indirection.