Add PORT envvar to all app pods
We set the exposed containerPort to .Values.appImage.appPort
so this ensures that we set the PORT envvar (used by our rails
apps pretty consistently) to the same value.

Currently government-frontend is failing its healthcheck as the
Dockerfile [1] specifies PORT=3090 so the app is not running
on the exposed port.

[1]: https://github.com/alphagov/government-frontend/blob/main/Dockerfile#L32