Use /healthcheck/live for readiness probe
The /healthcheck/ready endpoints check the connectivity to
and the health of dependencies such as caches, databases,
APIs, and external services.

This is inappropriate for a readiness probe in Kubernetes, which
should only check whether the app is ready to receive traffic,
(i.e. is Puma/Unicorn ready).

As an example, should memcached or content-store go down, this would cause
a cascading failure as some frontend apps would fail their healthchecks.

Instead the status checks should be exported as Prometheus metrics,
which will enable us to debug issues, but not prevent the app from
degrading gracefully (as some apps do) or serving requests which don't
require the external depencency (some apps have these).

This change will also make it easier to debug in Kubernetes, since
we'll be able to get our apps running and not have the pods
killed and restarted while we're debugging.