Set some default resource requests and limits for apps.
These are just an initial guesstimate that's intended to be better than
not setting any resource requests at all. It's highly likely we'll need
to adjust these when we start load testing, if not before.

We currently don't have anything that's using more than 0.5 GiB so I'm
not expecting this to cause any immediate issues for testing. If it
does, we can bump the memory a bit.