Disallow all robots except in production.
Invert the conditional so that we provide the disallow-all robots.txt
everywhere except in production. We wouldn't want crawlers to index a
non-production site if it were to be accidentally made public.

Also move the prod robots.txt into a separate file to make it easier to
read and maintain. Since the file doesn't vary by environment, include
it verbatim using `$.Files.Get` to further simplify matters.

Tested: inspected output of `helm template .` and
`helm template . --set govukEnvironment=production`.