Use `sh` rather than `bash` for assets-upload-job.
This makes it compatible with a wider range of base images. For example,
it makes it fail in a more helpful way when unintentionally run against
Router, which is currently based on Alpine.

Tested: ran the modified command via `kubectl exec` and checked the
success and failure (nonexistent directory) cases.