Disable service links (`*_SERVICE_*` env vars) on app pods.
Disable [service links] (the automatically-set environment variables
with names like `$service_SERVICE_{HOST,PORT}`) in the `govuk-rails-app`
chart. We're not using them and they create a lot of clutter in the
environment which can be a pain when debugging.

Maybe someday we could use them in Router and Plek, but that's unlikely
to happen soon.

[service links]: https://kubernetes.io/docs/concepts/services-networking/connect-applications-service/#environment-variables