Split out image tags into dedicated files
This creates a new directory image-tags/ which contains files which
specify the image tag to use in a given environment.

For instance:

```
├── Chart.yaml
├── image-tags
│   ├── integration
│   │   ├── asset-manager
│   │   ├── authenticating-proxy
│   │   ├── collections
```

To find the image that should be deployed to production for asset-manager,
one can read image-tags/production/asset-manager to find the appropriate
image tag.

The ArgoCD Application resources pull in the files during templating,
so any update of these files will trigger deployments to kubernetes.

This will make it easier to update the image tag for an Application.
In order to continuous deployment and release promotions automatically
we must only overwrite a single file.
It will also be easier to monitor releases using e.g. git log on image-tags/integration/asset-manager.

This approach introduces many new files, but overall the config isn't
more complicated. Originally I wanted to move all Helm values for an
app to dedicated file, since eventually values-<env>.yaml files will
be gigantic, hard to test, have a massive blast radius, and make it
harder to avoid duplication. However, I've avoided that for now, since
we probably want to take a step back and reconsider whether we want a
layer on top of or instead of Helm that would be a better fit.