Add the www hostname to the Router ingress.
Fastly appears to forward the Host header by default and we had trouble
trying to override it in VCL, so let's just try having the load balancer
accept the www hostname in addition to the www-origin one.

On the existing EC2 stack, the load balancer forwards everything to
Router regardless of the hostname.