Add pre-push hook to prevent pushing to main
We are no longer enforcing branch protection constraints on
administrators on this repository. This enables repo admins to
push to main. To prevent this, we have installed a git pre-push
hook.

The reason we have disabled this branch protection is we would like
the administrator user govuk-ci to bypass the PR and status check process
and push directly to main (automatically) without needing to raise a PR.