Serve assets from the existing (EC2) assets service for now.
Since we're focusing on the www frontend for now and it'll be a while
longer before we have Asset Manager fully functional in k8s, use the
existing Asset Manager domain for serving assets.

"Assets" here means attachements in GOV.UK content (like PDFs and
some images), not CSS/JS/images that are part of the GOV.UK apps.