Fix assets paths for Whitehall.
The assets upload job was defaulting to nonexistent paths for the three
Whitehall deployments. Since whitehall-admin, whitehall-frontend and
draft-whitehall-frontend each run the same image (though not necessarily
the same version) with a different .Release.Name and none of those names
match the assets directory inside the image, we need to override it.

We still run the assets upload job independently for each of the three
deployments because it's possible for them to be running different
versions and therefore need to upload different assets files - even
though the vast majority of the time it'll just be fruitlessly
uploading the same thing three times.