Vendor schemas for CRDs
In order to validate Kubernetes manifests against schemas,
we need to acquire those schemas. For speed we've cached them
here.

The expectation is that when new CRDs are adopted we will vendor
their schemas.

Vendoring the schemas is easy. Follow the documentation
on kubeconform README:
https://github.com/yannh/kubeconform#converting-an-openapi-file-to-a-json-schema
i.e. ./scripts/openapi2jsonschema.py <crd-manifest-url>
and mv the JSON schema file to the schemas directory.

This will enable us to validate manifests before merging.