Add kubeconform to CI GitHub Action
This adopts [kubeconform] for validating Kubernetes manifests
and also runs helm template for each chart in the charts directory.

[kubeconform]: https://github.com/yannh/kubeconform

Using helm template with the default values.yaml files
in charts is suboptimal - it would be better to test
the charts with values that will be provided when installed into
the cluster (rather than default/dummy values).

This at least validates that the Helm charts themselves can produce
valid Kubernetes manifests.