Fix request routing to Search API and Account API.
We intend to use Search API and Account API in EC2 for now, while we're
focusing on frontend (www) functionality. There were various
configuration issues preventing these requests from being routed
correctly.

- Configure Router to forward to EC2 search-api and account-api.
- Configure apps (via Plek) to use EC2 search-api.
- Use the internal load balancer addresses (govuk-internal.digital)
  where we can.
- It's `PLEK_SERVICE_ACCOUNT_API_URI`, not
  `PLEK_SERVICE_ACCOUNT_URI`, because the client code calls
  `Plek.find("account-api")`[1].

Not making these changes in the test environment because the EC2 stuff
in the test account is too broken for this to be meaningful there.

[1]: https://github.com/alphagov/gds-api-adapters/blob/599e216/lib/gds_api.rb#L34