Add promote release step to postsync workflow
This adds a workflowTemplate to update the image tag value in the helm
chart, which is used by the promote release step of the postsync
workflow. The release promotion will only occur on successful run of
Smokey, as to only promote non-broken releases. The script will also
only promote images where the image tag corrosponds to the latest commit
on main.  This ensure that automatic deployments only occur when try to
release the latest version, rather than a manual deployment.