Use the existing value for the EC2 domain in bootstrap-signon-resources.
Remove govukPublishingDomain (which was specific to this chart and
didn't include the environment, which is no good because the prod domain
follows a different pattern than the integration and staging one) and
replace it with publishingServiceDomainSuffix, which is already passed
into every chart via the argocd-apps chart and represents the whole
suffix including the environment.

While we're there, remove the unnecessary clusterDomain value, since we
can just specify a relative domain name and simplify the config at the
same time.