Specify correct URLs for EC2 apps in bootstrap-signon-resources.
The bootstrap script currently ignores the `home_uri` and `redirect_uri`
fields for Application objects which already exist, but it's super
misleading for the EC2 ones to have EKS URLs. Set them to the correct
URLs so that it's clearer that these are EC2 apps and so that we won't
break them if/when we improve the bootstrap script to make it update
these fields for existing objects.