Point Plek env vars at EC2 backends/APIs.
Add the vars for Link Checker API and Asset Manager since we're now
using those. (It's likely that they're only used in rake tasks on
frontend apps rather than as part of processing a request, but it's
easier to just set them anyway for now rather than trawl through the
code trying to prove a negative.)

Switch publishing-api to EC2 for now. (Again, not sure whether it's
really in the request path, but it's the simpler/safer option.)

Keep using the EKS Signon for now, since it's only relevant to the
publishing apps and draft stack, which aren't involved in milestones
1-4.