Improve reliability of update-image-tag script
This sets several shell options to improve the predictability of the
script. Adds a pipefile and the -e to exit the script immediately on a
failed command. Adds the -u to exit the script if a used variable is
unset. Adds -o to output the commands that are run to help debug the
script.