Disable 304 support for error pages
When S3 sends the stored error pages response, it has the Etag
and last-modified headers set. This causes Fastly to issue
a request with header `if-none-match` or `if-last-modified`
after the object expires in the fastly node.

This fastly behavior is not compatible with the `if` block
which returns a not found if URL is mixed case. Therefore,
we disables the 304 functionality in fastly for error pages
by removing the Etag and last-modified headers.

Refs:
1. [fastly doc](https://support.fastly.com/hc/en-us/community/posts/360041815632-304-Response-at-Fastly)
2. [smokey test](https://github.com/alphagov/smokey/blob/db2af4e5803d167f006fabe7ca204f52a511620a/features/gov_uk.feature#L28)