Be bug-compatible with the naming of `licencefinder`.
Also known as `licence-finder`. The repo name has the dash but the
domain name omits the dash. The assets path also omits the dash, as does
the Router backend name.

This should fix Router's inability to forward traffic to Licence Finder.

I'm leaving the name of the Signon app / ApiUser as `licence-finder`,
because it's easier that way and it's already inconsistent whether we
like it or not, so 🤷

This is the exact same problem and solution as for
`smartanswers`/`smart-answers`.

https://trello.com/c/F7vSHnws/841