Make MongoDB connection strings match cluster nodenames.
When a Mongo client connect to a cluster (replicaset), the client
library connects to one or more of the hosts given in the connection
string and requests list of nodes. It then tries to connect to the
hostnames in that list. The client libraries that we use seem to remove
the original hostnames (the ones in the connection string) from the list
of nodes that they'll connect to in future.  In some cases, this leads
to the client library being unable to reconnect because it's discarded
all the reachable hosts from the connection string and replaced them
with unreachable/unresolveable hosts that the cluster told it about.

The `mongoid` client library that we use in our Ruby apps seem to be
more brittle regarding this issue than the `mgo` library in Router, but
in both cases it's a good idea for stability to have the hostnames in
the connection string match what the cluster returns.

Annoyingly, the Mongo clusters are not configured consistently in this
regard: the Router one (on router_backend machines) has short hostnames
as nodenames whereas the other Mongo cluster uses FQDNs (without the
"stackname" component, i.e `blue`).

We include the stackname where specifying the additional search suffix
for Router, because we might as well save a DNS query. The `blue` thing
isn't going to change in the remaining lifetime of the GOV.UK EC2 stack.